move colors out of gl, start getters and setters
