+++ include/polyscope/colors.h
@@ -2,7 +2,6 @@
 #pragma once
 
 namespace polyscope {
-namespace gl {
 
 const glm::vec3 RGB_TEAL = {0., 178. / 255., 178. / 255.};
 const glm::vec3 RGB_BLUE = {150. / 255., 154. / 255., 255. / 255.};
@@ -16,5 +15,4 @@ const glm::vec3 RGB_LIGHTGRAY = {.8, .8, .8};
 const glm::vec3 RGB_DARKRED = {.2, .0, .0};
 const glm::vec3 RGB_PINK = {249. / 255., 45. / 255., 94. / 255.};
 
-} // namespace gl
 } // namespace polyscope
